[Filesystem] Moving Filesystem classes to a new Manager namespace for better organization.
