+++ src/Valkyrja/Application/Env.php
@@ -37,9 +37,9 @@ use Valkyrja\Cli\Middleware\Contract\InputReceivedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\ThrowableCaughtMiddleware;
 use Valkyrja\Crypt\Contract\Crypt;
 use Valkyrja\Crypt\SodiumCrypt;
-use Valkyrja\Filesystem\Contract\Filesystem;
-use Valkyrja\Filesystem\FlysystemFilesystem;
-use Valkyrja\Filesystem\LocalFlysystemFilesystem;
+use Valkyrja\Filesystem\Manager\Contract\Filesystem;
+use Valkyrja\Filesystem\Manager\FlysystemFilesystem;
+use Valkyrja\Filesystem\Manager\LocalFlysystemFilesystem;
 use Valkyrja\Http\Client\Contract\Client;
 use Valkyrja\Http\Client\GuzzleClient;
 use Valkyrja\Http\Message\Constant\HeaderName;
