+++ src/Valkyrja/Filesystem/Manager/Contract/Filesystem.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Filesystem\Contract;
+namespace Valkyrja\Filesystem\Manager\Contract;
 
 use Valkyrja\Filesystem\Enum\Visibility;
