+++ src/Valkyrja/Filesystem/Manager/InMemoryFilesystem.php
@@ -11,15 +11,15 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Filesystem;
+namespace Valkyrja\Filesystem\Manager;
 
 use Override;
 use Valkyrja\Exception\RuntimeException;
-use Valkyrja\Filesystem\Contract\Filesystem as Contract;
 use Valkyrja\Filesystem\Data\InMemoryFile;
 use Valkyrja\Filesystem\Data\InMemoryMetadata;
 use Valkyrja\Filesystem\Enum\Visibility;
 use Valkyrja\Filesystem\Exception\UnableToReadContentsException;
+use Valkyrja\Filesystem\Manager\Contract\Filesystem as Contract;
 
 use function fread;
 use function str_starts_with;
