+++ src/Valkyrja/Filesystem/Manager/S3FlysystemFilesystem.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Filesystem;
+namespace Valkyrja\Filesystem\Manager;
 
 /**
  * Class S3FlysystemFilesystem.
