+++ src/Valkyrja/Http/Middleware/Cache/CacheResponseMiddleware.php
@@ -16,8 +16,8 @@ namespace Valkyrja\Http\Middleware\Cache;
 use Override;
 use Throwable;
 use Valkyrja\Exception\RuntimeException;
-use Valkyrja\Filesystem\Contract\Filesystem;
-use Valkyrja\Filesystem\InMemoryFilesystem;
+use Valkyrja\Filesystem\Manager\Contract\Filesystem;
+use Valkyrja\Filesystem\Manager\InMemoryFilesystem;
 use Valkyrja\Http\Message\Enum\StatusCode;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Http\Message\Response\Contract\Response;
